[DOP-14063] Drop onetl._internal module
