+++ .github/workflows/data/core/tracked.txt
@@ -3,6 +3,5 @@ onetl/plugins/**
 onetl/impl/**
 onetl/hwm/**
 onetl/_util/**
-onetl/_internal.py
 onetl/log.py
 .github/workflows/data/core/**
