+++ onetl/connection/db_connection/kafka/kafka_ssl_protocol.py
@@ -10,8 +10,8 @@ try:
 except (ImportError, AttributeError):
     from pydantic import Field, SecretStr, validator  # type: ignore[no-redef, assignment]
 
-from onetl._internal import stringify
 from onetl._util.file import is_file_readable
+from onetl._util.spark import stringify
 from onetl.impl import GenericOptions, LocalPath
 
 if TYPE_CHECKING:
